requirements.txt
requests 
beautifulsoup4 
faiss-cpu 
sentence-transformers 
google-generativeai 
streamlit
python-dotenv

